feat: Replace hardcoded X-Company-ID header with extension function
The changes in this commit replace the hardcoded `X-Company-ID` header with a new extension function `withCompanyId()` that adds the header to the request. This makes the code more readable and maintainable, as the company ID handling is now centralized in a single location.

The changes were made to the following files:

- `DocumentApiImpl.kt`
- `TransactionFileApiImpl.kt`
- `TransactionExtractionApiImpl.kt`
- `TransactionApiImpl.kt`

The most important changes are the introduction of the `withCompanyId()` extension function and the replacement of the hardcoded header in all the API implementations.