feat(repository): add filtering options for transactions and documents
The changes in this commit add new extension functions to the `HttpRequestBuilder` class to support filtering options for transactions and documents. These new functions allow the caller to filter the results by:

- Document type
- Supplier ID
- Date range
- Amount range
- IDs

These filtering options are then used in the `TransactionApiImpl` and `DocumentApiImpl` classes to build the appropriate HTTP requests with the specified filters.