feat(onboarding): Add forget password click handler
The changes in this commit add a new click handler for the "Forgot password?" button on the login screen. This allows users to navigate to a separate screen to reset their password.

The changes include:

- Adding a new `onForgetPasswordClick` parameter to the `LoginScreen` composable function.
- Updating the `LoginScreen` implementation to call the new `onForgetPasswordClick` callback when the "Forgot password?" button is clicked.
- Adding a new `forgetPasswordScreen` variable to the `LoginScreen` to represent the navigation target for the forget password screen.
- Updating the `LoginScreenPreview` to include the new `onForgetPasswordClick` parameter.

These changes improve the user experience by providing a way for users to reset their password if they have forgotten it.