feat(company): add avatar field to Company model
The changes in this commit add a new `avatar` field to the `Company` model in the `shared/domain/src/commonMain/kotlin/ai/thepredict/domain/model/Common.kt` file. This field is used to store the URL of the company's avatar image.

Additionally, the `WorkspacesList` and `WorkspaceListItem` composables in the `application/ui/src/commonMain/kotlin/ai/thepredict/ui/Workspaces.kt` file have been updated to use the new `avatar` field instead of the old `url` field.

These changes are necessary to support the display of company avatars in the workspace list UI.