feat(onboarding): Separate full name into first and last name
The changes in this commit separate the full name field into first name and last name fields in the RegisterScreen. This allows for a more granular collection of user information during the registration process.

The key changes are:

- Rename the `fullName` variable to `firstName` and `lastName`
- Update the `PTextFieldName` components to use the new `firstName` and `lastName` variables
- Update the `createUser` function call in the `onRegisterClick` to pass the `firstName` and `lastName` separately

This change improves the user experience by providing a more intuitive and structured way for users to enter their name during registration.