feat: Add floating action button to navigation bar
The changes in this commit add a floating action button (FAB) to the navigation bar. The FAB is represented by a new `NavigationItem` called `AddDocument`, which has an icon and label. The `NavigationBar` component has been updated to accept a `fabItem` parameter, which is used to render the FAB. The `onItemClick` callback has also been added to handle clicks on the navigation items, including the FAB.

These changes are important as they provide a way for users to quickly access the "Add Document" functionality from the navigation bar, improving the overall user experience.